chore: some ideas around release builds not calling into metrics
