+++ com.unity.netcode.gameobjects/Runtime/Metrics/NetworkMetricsImpl.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b858c4c3c7e08884bafdba745e688dbb
+guid: dfaa1d7bcd4e56a41b0a75dbe89952c0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
